+++ lib/ox-ai-workers.rb
@@ -39,13 +39,14 @@ module OxAiWorkers
   class ConfigurationError < Error; end
 
   class Configuration
-    attr_accessor :model, :max_tokens, :temperature, :access_token
+    attr_accessor :model, :max_tokens, :temperature, :access_token, :auto_execute
 
     def initialize
       @access_token = nil
       @model = DEFAULT_MODEL
       @max_tokens = DEFAULT_MAX_TOKEN
       @temperature = DEFAULT_TEMPERATURE
+      @auto_execute = true
 
       [Array, NilClass, String, Symbol, Hash].each do |c|
         c.send(:include, OxAiWorkers::PresentCompat) unless c.method_defined?(:present?)